Отрисуй меня полностью - правки
