+++ js/utils.js
@@ -31,7 +31,7 @@ function getRandomArray(array) {
   return array.slice(0,getRandomIntInclusive(1, array.length-1));
 }
 
-const NEAR_ADS_COUNT = 1;
+const NEAR_ADS_COUNT = 10;
 
 const getAuthor = () => {
   return {
@@ -71,6 +71,10 @@ const getAd = () => {
   };
 };
 
-const getNearbyAds = new Array(NEAR_ADS_COUNT).fill(null).map(() => getAd());
 
-export {getNearbyAds};
+const createAbs = () => {
+  return new Array(NEAR_ADS_COUNT).fill(null).map(() => getAd());
+}
+
+
+export {createAbs};
